tech_research_agent/requirements.txt
streamlit
groq
qdrant-client
ollama
requests
python-dotenv
uuid